user_service/requirements.txt
nameko==2.14.1
kombu==5.2.4
bcrypt==4.1.3
PyJWT==2.8.0
requests==2.25.1
nameko-redis==1.0.0